Outsource parameter replacement helper to be used by default and tool modules
