+++ client/src/components/Workflow/Run/WorkflowRunDefaultStep.vue
@@ -16,10 +16,10 @@
 </template>
 
 <script>
-import { visitInputs } from "components/Form/utilities";
 import WorkflowIcons from "components/Workflow/icons";
 import FormDisplay from "components/Form/FormDisplay";
 import FormCard from "components/Form/FormCard";
+import { getReplacements } from "./model";
 
 export default {
     components: {
@@ -77,35 +77,12 @@ export default {
             }
         },
         wpData() {
-            this.onReplaceParams();
+            this.replaceParams = getReplacements(this.model.inputs, this.wpData);
         },
     },
     methods: {
         onChange(data) {
             this.$emit("onChange", this.model.index, data);
-        },
-        onReplaceParams() {
-            const params = {};
-            visitInputs(this.model.inputs, (input, name) => {
-                params[name] = input;
-            });
-            this.replaceParams = {};
-            _.each(params, (input, name) => {
-                if (input.wp_linked) {
-                    let newValue = input.value;
-                    const re = /\$\{(.+?)\}/g;
-                    let match;
-                    while ((match = re.exec(input.value))) {
-                        const wpValue = this.wpData[match[1]];
-                        if (wpValue) {
-                            newValue = newValue.split(match[0]).join(wpValue);
-                        }
-                    }
-                    if (newValue !== undefined) {
-                        this.replaceParams[name] = newValue;
-                    }
-                }
-            });
         },
         onValidation(validation) {
             this.$emit("onValidation", this.model.index, validation);
